requirements.txt
bottle # webserver
jsonpickle # to send complex datatypes
numpy
pandas
pyyaml